Send the full comment object to `get_avatar()` rather than just the email address.
`get_avatar()` can accept a user ID, email address, or comment object.  It's best to use a comment object because it provides the most context to plugins which use the `get_avatar` filter to change the response of the `get_avatar()` function.

Core WordPress always sends the full comment object to `get_avatar()` when rendering comment avatars.